+++ java/com/turbovnc/vncviewer/OptionsDialog.java
@@ -1,6 +1,6 @@
 /* Copyright (C) 2002-2005 RealVNC Ltd.  All Rights Reserved.
  * Copyright (C) 2011-2013 Brian P. Hinz
- * Copyright (C) 2012-2016 D. R. Commander.  All Rights Reserved.
+ * Copyright (C) 2012-2017 D. R. Commander.  All Rights Reserved.
  *
  * This is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -303,10 +303,15 @@ class OptionsDialog extends Dialog implements ActionListener, ChangeListener,
 
     Object[] spanOptions = {
       "Primary monitor only", "All monitors", "Automatic" };
-    JLabel spanLabel = new JLabel("Span mode:");
+    JLabel spanLabel;
+    if (VncViewer.isX11() && Viewport.isHelperAvailable())
+      spanLabel = new JLabel("Full-screen span mode:");
+    else
+      spanLabel = new JLabel("Span mode:");
     span = new JComboBox(spanOptions);
     span.addItemListener(this);
-    if (VncViewer.embed.getValue() || VncViewer.isX11()) {
+    if (VncViewer.embed.getValue() ||
+        (VncViewer.isX11() && !Viewport.isHelperAvailable())) {
       spanLabel.setEnabled(false);
       span.setEnabled(false);
     }